fix: cast Canary attention mask to Q/K/V dtype for bf16 inference (#43)

+++ src/vemoizer/canary_mlx.py
@@ -234,7 +234,9 @@ class RelPosMultiHeadAttention(nn.Module):
 
         # relative bias: q_v · p^T, shifted so that position (i, j) sees (i - j)
         matrix_bd = self._rel_shift(q_v @ p.swapaxes(-2, -1))[:, :, :, :q_seq]
-        matrix_bd = matrix_bd * self.scale
+        # mx.fast.scaled_dot_product_attention requires the mask dtype to
+        # promote to the Q/K/V output dtype, so cast to q's dtype.
+        matrix_bd = (matrix_bd * self.scale).astype(q.dtype)
         # mx.fast.scaled_dot_product_attention expects a mask of at most rank 4,
         # so fold (batch, head) into a single axis: (b,1,h,q,p)->(b*h,q,p).
         matrix_bd = mx.expand_dims(matrix_bd, 1).reshape(
